Set tabstop only.
I used to set tabstop, softtabstop and shiftwidth because I didn't know how they
related to eachother.

Now I know sts is how it "looks" (and will default to ts), and that shiftwidth
will account for retabbing and so on.

Unset them all so that there is a single source of truth (tabstop).